+++ packages/provider/src/api/verify.ts
@@ -127,6 +127,8 @@ export function prosopoVerifyRouter(env: ProviderEnvironment): Router {
 						userAccessRulesStorage,
 						email,
 						clientRecord.settings.spamEmailDomainCheckEnabled,
+						clientRecord.settings.spamFilter,
+						clientRecord.settings.trafficFilter,
 					);
 
 				req.logger.debug(() => ({ data: { response } }));
@@ -239,6 +241,8 @@ export function prosopoVerifyRouter(env: ProviderEnvironment): Router {
 						userAccessRulesStorage,
 						email,
 						clientRecord.settings.spamEmailDomainCheckEnabled,
+						clientRecord.settings.spamFilter,
+						clientRecord.settings.trafficFilter,
 					);
 
 				const verificationResponse: VerificationResponse =