+++ dev/scripts/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prosopo/scripts
 
+## 3.1.74
+### Patch Changes
+
+- Updated dependencies [fc514dd]
+- Updated dependencies [42650db]
+  - @prosopo/provider@3.23.0
+  - @prosopo/locale@3.2.0
+  - @prosopo/types@3.14.0
+  - @prosopo/types-env@2.9.0
+  - @prosopo/env@3.4.0
+  - @prosopo/cli@3.6.10
+  - @prosopo/common@3.1.32
+  - @prosopo/datasets@3.1.14
+  - @prosopo/keyring@2.9.20
+  - @prosopo/dotenv@3.0.37
+
 ## 3.1.73
 ### Patch Changes