+++ packages/api-route/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @prosopo/api-route
 
+## 2.6.40
+### Patch Changes
+
+  - @prosopo/common@3.1.32
+
 ## 2.6.39
 ### Patch Changes