+++ packages/detector/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prosopo/detector
 
+## 3.4.23
+### Patch Changes
+
+- Updated dependencies [fc514dd]
+- Updated dependencies [42650db]
+  - @prosopo/types@3.14.0
+
 ## 3.4.22
 ### Patch Changes