+++ packages/load-balancer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @prosopo/load-balancer
 
+## 2.8.37
+### Patch Changes
+
+- Updated dependencies [fc514dd]
+- Updated dependencies [42650db]
+  - @prosopo/types@3.14.0
+  - @prosopo/common@3.1.32
+
 ## 2.8.36
 ### Patch Changes