+++ packages/server/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prosopo/server
 
+## 2.9.72
+### Patch Changes
+
+- Updated dependencies [fc514dd]
+- Updated dependencies [7be39c4]
+- Updated dependencies [42650db]
+  - @prosopo/types@3.14.0
+  - @prosopo/api@3.2.10
+  - @prosopo/common@3.1.32
+  - @prosopo/keyring@2.9.20
+  - @prosopo/load-balancer@2.8.37
+
 ## 2.9.71
 ### Patch Changes