+++ packages/types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prosopo/types
 
+## 3.14.0
+### Minor Changes
+
+- 42650db: Add better spam rules and move ipinfo service to local instead of external
+
+### Patch Changes
+
+- fc514dd: ability to block different types of traffic
+- Updated dependencies [fc514dd]
+- Updated dependencies [42650db]
+  - @prosopo/locale@3.2.0
+
 ## 3.13.3
 ### Patch Changes