+++ packages/user-access-policy/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prosopo/user-access-policy
 
+## 3.6.21
+### Patch Changes
+
+- Updated dependencies [fc514dd]
+- Updated dependencies [7be39c4]
+- Updated dependencies [42650db]
+  - @prosopo/types@3.14.0
+  - @prosopo/api@3.2.10
+  - @prosopo/common@3.1.32
+  - @prosopo/api-route@2.6.40
+  - @prosopo/redis-client@1.0.17
+
 ## 3.6.20
 ### Patch Changes